conversion to giter8 template format
