--- activator.properties
@@ -1,5 +1,0 @@
-name=minimal-scalismo-seed
-title=Statistical Shape Modelling in Scala
-description=A starter project
-tags=seed,shape-modelling,medical-imaging,machine-learning,scala
-authorLogo=http://urz.unibas.ch/files/documents/unilogoschwarz.gif
