--- build.sbt
@@ -1,32 +1,0 @@
-organization  := "ch.unibas.cs.gravis"
-
-name := """minimal-scalismo-seed"""
-version       := "0.6"
-
-scalaVersion  := "2.12.6"
-
-scalacOptions := Seq("-unchecked", "-deprecation", "-encoding", "utf8")
-
-resolvers += Resolver.bintrayRepo("unibas-gravis", "maven")
-
-resolvers += Opts.resolver.sonatypeSnapshots
-
-
-libraryDependencies  ++= Seq(
-            "ch.unibas.cs.gravis" % "scalismo-native-all" % "4.0.+",
-            "ch.unibas.cs.gravis" %% "scalismo-ui" % "0.13.1"
-)
-
-assemblyJarName in assembly := "executable.jar"
-
-mainClass in assembly := Some("com.example.ExampleApp")
-
-
-assemblyMergeStrategy in assembly <<= (assemblyMergeStrategy in assembly) { (old) =>
-  {
-    case PathList("META-INF", "MANIFEST.MF") => MergeStrategy.discard
-    case PathList("META-INF", s) if s.endsWith(".SF") || s.endsWith(".DSA") || s.endsWith(".RSA") => MergeStrategy.discard
-    case "reference.conf" => MergeStrategy.concat
-    case _ => MergeStrategy.first
-  }
-}
