--- data/README
@@ -1,5 +1,0 @@
-The provided example data are part of the Basel Face model, which is available at
-http://faces.cs.unibas.ch/bfm/.
-
-Please do not redistribute them without the consent of the University of Basel. 
-Copyright 2009, University of Basel
