--- project/plugins.sbt
@@ -1,2 +1,0 @@
-addSbtPlugin("com.typesafe.sbteclipse" % "sbteclipse-plugin" % "4.0.0")
-addSbtPlugin("com.eed3si9n" % "sbt-assembly" % "0.14.3")
