+++ src/main/g8/.gitignore
@@ -1,0 +1,1 @@
+../../../target
