+++ src/main/g8/default.properties
@@ -1,0 +1,3 @@
+name=ScalismoSeed
+description=A minimal Scalismo project
+verbatim=*.ply
